==> notebooks/1_prepare_documents_to_improve.ipynb <==
This notebook will run a C-SEO method on a list of documents from a domain of the benchmark.
We already provide the improved documents for all C-SEO methods in `data/{partition}/selected_docs.json`

import json
from data import Benchmark
import os
import random

notebook_dir = os.getcwd()
project_root = os.path.abspath(os.path.join(notebook_dir, ".."))

# Setup

data_path = "parameterlab/c-seo-bench"
list_splits = ["retail", "videogames", "news", "web", "debate", "books"]

doc_type_mapping = {
    "books": "Synopsis",
    "web": "Web Page Snippet",
    "debate": "Web Page Snippet",
    "news": "News Article",
    "retail": "Product Description",
    "videogames": "Game Description",
}

# Parameters

split_name = list_splits[0]

# loading the dataset
dataset = Benchmark(
    num_docs_in_context=10,
    method="baseline",
    sample_size=None,
    data_path=data_path,
    split=split_name,
    doc_type=doc_type_mapping[split_name],
)

dataset[0]

# pick the documents you want to improve using a c-seo method
query_idx2list_docs = dict()
for query_idx, query_point in enumerate(dataset):
    # randomly pick a document to improve
    doc_indices = random.sample(
        range(len(query_point["list_docs"])), k=1
    )  # 1 document per query. You can change k to pick more documents (i.e., multiple adopters)
    docs = dict()
    for doc_idx in doc_indices:
        doc = query_point["list_docs"][doc_idx]
        docs[doc_idx] = {"doc": doc}
    query_idx2list_docs[query_idx] = docs

# save the selected documents to a file
output_folder = os.path.join(project_root, "data", split_name)
if not os.path.exists(output_folder):
    os.makedirs(output_folder)
with open(os.path.join(output_folder, "selected_docs.json"), "w") as f:
    json.dump(query_idx2list_docs, f, indent=4)
